Left Encoder Bugged
Left encoder doesn't register a HIGH signal with the STM. Not sure if
this is a software or hardware problem. A deeper level of debugging will
be required. For now the code has the left encoder structure and call
functions. For now I will be moving on to the ADC module to keep on
schedule.